Refactor: Game Socket Move
