+++ src/App.tsx
@@ -3,7 +3,6 @@ import AdminMain from "./pages/Main/user";
 import Login from "./pages/Login";
 import Example from "./pages/Example";
 import Game from "./pages/Game";
-import Keyword from "./components/Game/Keyword";
 import GameLists from "./components/Main/GameLists";
 
 const App = () => {
@@ -14,7 +13,6 @@ const App = () => {
         <Route path="/main" element={<AdminMain />} />
         <Route path="/game" element={<Game />} />
         <Route path="/example" element={<Example />} />
-        <Route path="/keyword" element={<Keyword />} />
         <Route path="/gameLists" element={<GameLists />} />
       </Routes>
     </>
