+++ src/data/category.json
@@ -28,8 +28,5 @@
       "category": "사물",
       "keyword": ["가위", "분필", "종이컵", "양말", "마우스", "거울"]
     }
-  ],
-  "users": {
-    "name": ["용훈", "수연", "혜민", "정아", "연수"]
-  }
+  ]
 }
